+++ documentation/docs/changelog.md
@@ -2,6 +2,8 @@
 
 ## yyyy.mm.dd - version 1.5.0
 
+* Allow commands to register [argument validators](@TODO: link to the documentation <2024.06.12>)
+
 ## 2024.06.11 - version 1.4.0
 
 * Add a factory to create item instances from multiple sources