+++ documentation/docs/changelog.md
@@ -2,8 +2,10 @@
 
 ## yyyy.mm.dd - version 1.5.0
 
-* Allow commands to register [argument validators](resources/commands/command#validating-arguments)
-* Default to the [help operation](resources/commands/commands-handler#the-help-operation)
+* Allow commands to register
+[argument validators](resources/commands/command#validating-arguments)
+* Default to the
+[help operation](resources/commands/commands-handler#the-help-operation)
 when it's missing in the addon main slash command
 * Watch for the [PLAYER_LEVEL_UP](resources/facades/events#player_level_up) 
 event and forward it to the library listeners